// apps/api/src/database/constraints/codes.ts
export enum ConstraintCode {
	// Foreign key violations
	CUSTOMER_TYPE_NOT_FOUND = "CUSTOMER_TYPE_NOT_FOUND",
	AGENCY_NOT_FOUND = "AGENCY_NOT_FOUND",
	USER_NOT_FOUND = "USER_NOT_FOUND",
	ROLE_NOT_FOUND = "ROLE_NOT_FOUND",
	CUSTOMER_NOT_FOUND = "CUSTOMER_NOT_FOUND",
	GOODS_TYPE_NOT_FOUND = "GOODS_TYPE_NOT_FOUND",
	GOODS_NOT_FOUND = "GOODS_NOT_FOUND",
	STATION_NOT_FOUND = "STATION_NOT_FOUND",
	PORT_NOT_FOUND = "PORT_NOT_FOUND",
	BERTH_NOT_FOUND = "BERTH_NOT_FOUND",
	SIDING_NOT_FOUND = "SIDING_NOT_FOUND",
	ORDER_NOT_FOUND = "ORDER_NOT_FOUND",
	ORDER_STATUS_NOT_FOUND = "ORDER_STATUS_NOT_FOUND",
	MOVEMENT_TYPE_NOT_FOUND = "MOVEMENT_TYPE_NOT_FOUND",
	UNIT_NOT_FOUND = "UNIT_NOT_FOUND",
	PICKUP_LOCATION_TYPE_NOT_FOUND = "PICKUP_LOCATION_TYPE_NOT_FOUND",
	DISPATCH_TYPE_NOT_FOUND = "DISPATCH_TYPE_NOT_FOUND",
	PROGRAM_STATUS_NOT_FOUND = "PROGRAM_STATUS_NOT_FOUND",
	FORECAST_PROGRAM_NOT_FOUND = "FORECAST_PROGRAM_NOT_FOUND",
	ATTRIBUTE_NOT_FOUND = "ATTRIBUTE_NOT_FOUND",
	ACCESSORY_OPERATION_NOT_FOUND = "ACCESSORY_OPERATION_NOT_FOUND",
	REJECTION_REASON_NOT_FOUND = "REJECTION_REASON_NOT_FOUND",
	CLAIM_NOT_FOUND = "CLAIM_NOT_FOUND",
	CLAIM_TYPE_NOT_FOUND = "CLAIM_TYPE_NOT_FOUND",
	CLAIM_STATUS_NOT_FOUND = "CLAIM_STATUS_NOT_FOUND",
	TRAIN_NOT_FOUND = "TRAIN_NOT_FOUND",
	WAGON_NOT_FOUND = "WAGON_NOT_FOUND",
	NOTIFICATION_TYPE_NOT_FOUND = "NOTIFICATION_TYPE_NOT_FOUND",
	NOTIFICATION_CHANNEL_NOT_FOUND = "NOTIFICATION_CHANNEL_NOT_FOUND",
	DTM_REQUEST_TYPE_NOT_FOUND = "DTM_REQUEST_TYPE_NOT_FOUND",
	PARENT_ORDER_NOT_FOUND = "PARENT_ORDER_NOT_FOUND",
	PERMISSION_NOT_FOUND = "PERMISSION_NOT_FOUND",

	// Unique violations
	DUPLICATE_AGENCY_NAME = "DUPLICATE_AGENCY_NAME",
	DUPLICATE_USER_EMAIL = "DUPLICATE_USER_EMAIL",
	DUPLICATE_USER_EMPLOYEE_CODE = "DUPLICATE_USER_EMPLOYEE_CODE",
	DUPLICATE_CUSTOMER_CODE = "DUPLICATE_CUSTOMER_CODE",
	DUPLICATE_CUSTOMER_TYPE_NAME = "DUPLICATE_CUSTOMER_TYPE_NAME",
	DUPLICATE_ROLE_NAME = "DUPLICATE_ROLE_NAME",
	DUPLICATE_PERMISSION_NAME = "DUPLICATE_PERMISSION_NAME",
	DUPLICATE_GOODS_TYPE_NAME = "DUPLICATE_GOODS_TYPE_NAME",
	DUPLICATE_GOODS_CODE = "DUPLICATE_GOODS_CODE",
	DUPLICATE_STATION_NAME = "DUPLICATE_STATION_NAME",
	DUPLICATE_STATION_CODE = "DUPLICATE_STATION_CODE",
	DUPLICATE_PORT_NAME = "DUPLICATE_PORT_NAME",
	DUPLICATE_SIDING_NAME = "DUPLICATE_SIDING_NAME",
	DUPLICATE_ATTRIBUTE_NAME = "DUPLICATE_ATTRIBUTE_NAME",
	DUPLICATE_ORDER_NUMBER = "DUPLICATE_ORDER_NUMBER",
	DUPLICATE_ORDER_STATUS_NAME = "DUPLICATE_ORDER_STATUS_NAME",
	DUPLICATE_MOVEMENT_TYPE_NAME = "DUPLICATE_MOVEMENT_TYPE_NAME",
	DUPLICATE_UNIT_NAME = "DUPLICATE_UNIT_NAME",
	DUPLICATE_PICKUP_LOCATION_TYPE_NAME = "DUPLICATE_PICKUP_LOCATION_TYPE_NAME",
	DUPLICATE_DISPATCH_TYPE_NAME = "DUPLICATE_DISPATCH_TYPE_NAME",
	DUPLICATE_PROGRAM_STATUS_NAME = "DUPLICATE_PROGRAM_STATUS_NAME",
	DUPLICATE_PROGRAM_NUMBER = "DUPLICATE_PROGRAM_NUMBER",
	DUPLICATE_ACCESSORY_OPERATION_NAME = "DUPLICATE_ACCESSORY_OPERATION_NAME",
	DUPLICATE_REJECTION_REASON_NAME = "DUPLICATE_REJECTION_REASON_NAME",
	DUPLICATE_SHIPPING_COMPANY_NAME = "DUPLICATE_SHIPPING_COMPANY_NAME",
	DUPLICATE_VESSEL_NAME = "DUPLICATE_VESSEL_NAME",
	DUPLICATE_IMPORTER_NAME = "DUPLICATE_IMPORTER_NAME",
	DUPLICATE_REPRESENTATIVE_NAME = "DUPLICATE_REPRESENTATIVE_NAME",
	DUPLICATE_CLAIM_TYPE_NAME = "DUPLICATE_CLAIM_TYPE_NAME",
	DUPLICATE_CLAIM_STATUS_NAME = "DUPLICATE_CLAIM_STATUS_NAME",
	DUPLICATE_TRAIN_NUMBER = "DUPLICATE_TRAIN_NUMBER",
	DUPLICATE_TRAIN_EXTERNAL_ID = "DUPLICATE_TRAIN_EXTERNAL_ID",
	DUPLICATE_WAGON_NUMBER = "DUPLICATE_WAGON_NUMBER",
	DUPLICATE_WAGON_EXTERNAL_ID = "DUPLICATE_WAGON_EXTERNAL_ID",
	DUPLICATE_NOTIFICATION_TYPE_NAME = "DUPLICATE_NOTIFICATION_TYPE_NAME",
	DUPLICATE_NOTIFICATION_CHANNEL_NAME = "DUPLICATE_NOTIFICATION_CHANNEL_NAME",
	DUPLICATE_DTM_REQUEST_TYPE_NAME = "DUPLICATE_DTM_REQUEST_TYPE_NAME",
	DUPLICATE_PASSWORD_RESET_TOKEN = "DUPLICATE_PASSWORD_RESET_TOKEN",
	DUPLICATE_SESSION_TOKEN = "DUPLICATE_SESSION_TOKEN",

	// Composite unique violations
	DUPLICATE_ORDER_ATTRIBUTE = "DUPLICATE_ORDER_ATTRIBUTE",
	DUPLICATE_PARAMETRIZATION = "DUPLICATE_PARAMETRIZATION",
	DUPLICATE_CUSTOMER_PARAMETRIZATION = "DUPLICATE_CUSTOMER_PARAMETRIZATION",
	DUPLICATE_ORDER_SHARE = "DUPLICATE_ORDER_SHARE",
	DUPLICATE_PROGRAM_CONVOI = "DUPLICATE_PROGRAM_CONVOI",
}